Updated main to loop over several hyperparameters arrays. Updated pytorch model with fix on variable type. Undone changes in covid19 feature script

+++ pytorch_model.py
@@ -25,7 +25,7 @@ class Net(T.nn.Module):
     self.relu1 = T.nn.ReLU()
     
     # 2. Fully connect layer
-    hl2_size = hidden_size / 2
+    hl2_size = int(hidden_size / 2)
     self.fc2 = T.nn.Linear(hidden_size, hl2_size)
     self.relu2 = T.nn.ReLU()
